move sitemetadata to hooks folder
