+++ src/components/Footer.js
@@ -3,8 +3,8 @@ import React from 'react';
 import styled from 'styled-components';
 import { StyledHyperLink as SHL } from '../components/Shared';
 import ThemeSelect from '../components/ThemeSelect';
+import useSiteMetadata from '../hooks/siteMetadata';
 import { media } from '../theme/globalStyle';
-import useSiteMetadata from './SiteMetadata';
 
 // import { Dump } from '../utils/helpers'
